Fixed rain level trace + added comments
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@3163 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 15c17217aceb03c92e54150a5afb2528e9e312e5
Former-commit-id: cc17722e7b3ed5aacff6c4fdb1e13f2ccdeec4d2

+++ src/modules/graphic/ssggraph/grrain.cpp
@@ -87,7 +87,7 @@ cGrRain::~cGrRain(void)
 void cGrRain::initialize(int rain, float precipitationDensity) 
 {
 	GfLogInfo("Precipitation : Density = %d %%, initial rain strength = %d\n",
-			  precipitationDensity, rain);
+			  (int)precipitationDensity, rain);
 	precipitation_density = precipitationDensity;
 }
